On HP-UX, with aCC in 64bit mode, constants with the LL or ULL suffixes are silently changed to long and ulong. This fixes the problem with a static cast, and on most platforms the cast will probably be removed at compile time.
[git-p4: depot-paths = "//depot/qt/main/": change = 160543]

+++ src/core/global/qglobal.h
@@ -696,8 +696,8 @@ typedef unsigned long Q_ULONG;      // word up to 64 bit unsigned
 typedef __int64 Q_INT64;            // 64 bit signed
 typedef unsigned __int64 Q_UINT64;  // 64 bit unsigned
 #else
-#  define Q_INT64_C(c) c ## LL     // signed 64 bit constant
-#  define Q_UINT64_C(c) c ## ULL   // unsigned 64 bit constant
+#  define Q_INT64_C(c) static_cast<long long>(c ## LL)            // signed 64 bit constant
+#  define Q_UINT64_C(c) static_cast<unsigned long long>(c ## ULL) // unsigned 64 bit constant
 typedef long long Q_INT64;          // 64 bit signed
 typedef unsigned long long Q_UINT64;// 64 bit unsigned
 #endif
